# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MicroLink Control System \u2014 Data Platform API",
    "description": "**Stream B Data Platform API** \u2014 the central nervous system of MCS.\n\nThis is the contract that Stream C (Business Systems) and Stream D (Frontend)\ndevelop against. All telemetry, alarm, billing, and real-time data flows through\nthese endpoints.\n\n## Authentication\nAPI key via `X-API-Key` header. Multi-tenant isolation via row-level security\nscoped to the authenticated tenant.\n\n## Versioning\nAll endpoints are prefixed with `/api/v1`. Breaking changes will increment\nthe version number.\n\n## Rate Limits\n- REST endpoints: 1,000 req/min per API key\n- WebSocket connections: 10 per API key\n- Telemetry queries spanning >90 days: 10 req/min\n\n## Real-Time Feeds\nWebSocket endpoints provide live data streams backed by Redis pub/sub.\nReconnect with exponential backoff on disconnect.\n",
    "version": "0.1.0",
    "contact": {
      "name": "MicroLink MCS Team",
      "email": "engineering@microlinkdc.com"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000/api/v1",
      "description": "Local development"
    },
    {
      "url": "https://api.mcs.microlinkdc.com/api/v1",
      "description": "Production"
    }
  ],
...
  "paths": {
    "/sites": {
      "get": {
        "tags": [
          "Sites"
        ],
        "summary": "List all sites",
        "operationId": "listSites",
        "parameters": [
          {
            "$ref": "#/components/parameters/Offset"
          },
          {
            "$ref": "#/components/parameters/Limit"
          }
        ],
        "responses": {
          "200": {
            "description": "List of sites with block counts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Site"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/sites/{slug}": {
      "get": {
        "tags": [
          "Sites"
        ],
        "summary": "Get site details",
        "operationId": "getSite",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "baldwinsville"
          }
        ],
        "responses": {
          "200": {
            "description": "Site details with config",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteDetail"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/blocks": {
      "get": {
        "tags": [
          "Blocks"
        ],
        "summary": "List blocks",
        "operationId": "listBlocks",
        "parameters": [
          {
            "name": "site_slug",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Filter by site"
          },
          {
            "$ref": "#/components/parameters/Offset"
          },
          {
            "$ref": "#/components/parameters/Limit"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Block"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/blocks/{slug}": {
      "get": {
        "tags": [
          "Blocks"
        ],
        "summary": "Get block details",
        "operationId": "getBlock",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "block-01"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlockDetail"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/equipment/{block_slug}": {
      "get": {
        "tags": [
          "Blocks"
        ],
        "summary": "List equipment in a block",
        "operationId": "listEquipment",
        "parameters": [
          {
            "name": "block_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Equipment"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/sensors/{block_slug}": {
      "get": {
        "tags": [
          "Blocks"
        ],
        "summary": "List sensors in a block",
        "operationId": "listSensors",
        "parameters": [
          {
            "name": "block_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "subsystem",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "electrical",
                "thermal-l1",
                "thermal-l2",
                "thermal-l3",
                "thermal-reject",
                "thermal-safety",
                "environmental",
                "network",
                "security"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Sensor"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/telemetry": {
      "get": {
        "tags": [
          "Telemetry"
        ],
        "summary": "Query historical telemetry",
        "description": "Query telemetry for a single sensor over a time range.\n**Automatically selects the optimal aggregate tier** based on the time span.\nOverride with `agg` parameter for explicit control.\n\n| Time Range | Tier | Resolution |\n|-----------|------|------------|\n| < 2 hours | raw | 5 seconds |\n| 2h \u2013 12h | agg_1min | 1 minute |\n| 12h \u2013 3d | agg_5min | 5 minutes |\n| 3d \u2013 90d | agg_1hour | 1 hour |\n| > 90 days | agg_1day | 1 day |\n",
        "operationId": "queryTelemetry",
        "parameters": [
          {
            "name": "sensor_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Sensor ID"
          },
          {
            "$ref": "#/components/parameters/Start"
          },
          {
            "$ref": "#/components/parameters/End"
          },
          {
            "name": "agg",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "raw",
                "1min",
                "5min",
                "1hour",
                "1day"
              ]
            },
            "description": "Override aggregate tier"
          }
        ],
        "responses": {
          "200": {
            "description": "Telemetry data with metadata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TelemetryResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/telemetry/latest": {
      "get": {
        "tags": [
          "Telemetry"
        ],
        "summary": "Get latest value per sensor",
        "description": "Returns the most recent reading for every sensor in a block.\nUses DISTINCT ON for single-pass efficiency. Powers the NOC\ndashboard \"current state\" view.\n",
        "operationId": "getLatestValues",
        "parameters": [
          {
            "name": "block_slug",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "subsystem",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlockLatestResponse"
                }
              }
            }
          }
        }
      }
    },
    "/telemetry/multi": {
      "get": {
        "tags": [
          "Telemetry"
        ],
        "summary": "Query multiple sensors",
        "description": "Query up to 50 sensors in one call for comparison charts.",
        "operationId": "queryMultiSensor",
        "parameters": [
          {
            "name": "sensor_ids",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Comma-separated sensor IDs (max 50)",
            "example": "42,43,44,45"
          },
          {
            "$ref": "#/components/parameters/Start"
          },
          {
            "$ref": "#/components/parameters/End"
          },
          {
            "name": "agg",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "raw",
                "1min",
                "5min",
                "1hour",
                "1day"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MultiTelemetryResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "Offset": {
        "name": "offset",
        "in": "query",
        "schema": {
          "type": "integer",
          "default": 0,
          "minimum": 0
        }
      },
      "Limit": {
        "name": "limit",
        "in": "query",
        "schema": {
          "type": "integer",
          "default": 100,
          "minimum": 1,
          "maximum": 1000
        }
      },
      "Start": {
        "name": "start",
        "in": "query",
        "required": true,
        "schema": {
          "type": "string",
          "format": "date-time"
        },
        "description": "Start time (ISO 8601)"
      },
      "End": {
        "name": "end",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string",
          "format": "date-time"
        },
        "description": "End time (ISO 8601, default now)"
      }
    },
    "responses": {
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "detail": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "schemas": {
      "Site": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "slug": {
            "type": "string",
            "example": "baldwinsville"
          },
          "name": {
            "type": "string",
            "example": "Baldwinsville Brewery"
          },
          "region": {
            "type": "string",
            "nullable": true,
            "example": "US-East"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "commissioning",
              "decommissioned"
            ],
            "example": "active"
          },
          "latitude": {
            "type": "number",
            "nullable": true
          },
          "longitude": {
            "type": "number",
            "nullable": true
          },
          "block_count": {
            "type": "integer",
            "example": 2
          },
          "total_capacity_mw": {
            "type": "number",
            "example": 2.0
          }
        }
      },
      "SiteDetail": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Site"
          },
          {
            "type": "object",
            "properties": {
              "config": {
                "type": "object",
                "nullable": true
              },
              "created_at": {
                "type": "string",
                "format": "date-time",
                "nullable": true
              }
            }
          }
        ]
      },
      "Block": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "slug": {
            "type": "string",
            "example": "block-01"
          },
          "site_slug": {
            "type": "string"
          },
          "site_name": {
            "type": "string"
          },
          "capacity_mw": {
            "type": "number",
            "example": 1.0
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "commissioning",
              "standby",
              "decommissioned"
            ]
          },
          "commissioned_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "sensor_count": {
            "type": "integer",
            "example": 300
          }
        }
      },
      "BlockDetail": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Block"
          },
          {
            "type": "object",
            "properties": {
              "config": {
                "type": "object",
                "nullable": true
              },
              "equipment_count": {
                "type": "integer"
              }
            }
          }
        ]
      },
      "Equipment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "tag": {
            "type": "string",
            "example": "CDU-01"
          },
          "type": {
            "type": "string",
            "example": "coolant_distribution_unit"
          },
          "subsystem": {
            "type": "string",
            "enum": [
              "electrical",
              "thermal-l1",
              "thermal-l2",
              "thermal-l3",
              "thermal-reject",
              "thermal-safety",
              "environmental",
              "network",
              "security"
            ]
          },
          "block_slug": {
            "type": "string"
          },
          "metadata": {
            "type": "object",
            "nullable": true
          },
          "sensor_count": {
            "type": "integer"
          }
        }
      },
      "Sensor": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "tag": {
            "type": "string",
            "example": "CDU-01-T-SUP"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "unit": {
            "type": "string",
            "nullable": true,
            "example": "\u00b0C"
          },
          "subsystem": {
            "type": "string"
          },
          "equipment_tag": {
            "type": "string"
          },
          "block_slug": {
            "type": "string"
          },
          "range_min": {
            "type": "number",
            "nullable": true
          },
          "range_max": {
            "type": "number",
            "nullable": true
          },
          "poll_rate_ms": {
            "type": "integer",
            "nullable": true,
            "example": 5000
          },
          "alarm_thresholds": {
            "type": "object",
            "nullable": true,
            "description": "Keys: HH, H, L, LL. Values: {value, priority, delay_s}",
            "example": {
              "HH": {
                "value": 60.0,
                "priority": "P0",
                "delay_s": 0
              },
              "H": {
                "value": 55.0,
                "priority": "P2",
                "delay_s": 30
              }
            }
          }
        }
      },
      "TelemetryPoint": {
        "type": "object",
        "properties": {
          "bucket": {
            "type": "string",
            "format": "date-time"
          },
          "val_avg": {
            "type": "number",
            "example": 32.14
          },
          "val_min": {
            "type": "number",
            "example": 31.8
          },
          "val_max": {
            "type": "number",
            "example": 32.6
          },
          "val_last": {
            "type": "number",
            "example": 32.2
          },
          "sample_count": {
            "type": "integer",
            "example": 12
          },
          "quality_ratio": {
            "type": "number",
            "nullable": true,
            "example": 1.0,
            "description": "Fraction of GOOD samples (0.0\u20131.0)"
          }
        }
      },
      "TelemetryResponse": {
        "type": "object",
        "properties": {
          "sensor_id": {
            "type": "integer"
          },
          "sensor_tag": {
            "type": "string"
          },
          "unit": {
            "type": "string",
            "nullable": true
          },
          "tier": {
            "type": "string",
            "enum": [
              "telemetry",
              "agg_1min",
              "agg_5min",
              "agg_1hour",
              "agg_1day"
            ],
            "description": "Aggregate tier used"
          },
          "start": {
            "type": "string",
            "format": "date-time"
          },
          "end": {
            "type": "string",
            "format": "date-time"
          },
          "point_count": {
            "type": "integer"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TelemetryPoint"
            }
          }
        }
      },
      "LatestValue": {
        "type": "object",
        "properties": {
          "sensor_id": {
            "type": "integer"
          },
          "tag": {
            "type": "string"
          },
          "subsystem": {
            "type": "string"
          },
          "unit": {
            "type": "string",
            "nullable": true
          },
          "value": {
            "type": "number"
          },
          "quality": {
            "type": "integer",
            "description": "0=GOOD, 1=UNCERTAIN, 2=BAD"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "BlockLatestResponse": {
        "type": "object",
        "properties": {
          "block_slug": {
            "type": "string"
          },
          "sensor_count": {
            "type": "integer"
          },
          "readings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LatestValue"
            }
          }
        }
      },
      "MultiTelemetryResponse": {
        "type": "object",
        "properties": {
          "tier": {
            "type": "string"
          },
          "start": {
            "type": "string",
            "format": "date-time"
          },
          "end": {
            "type": "string",
            "format": "date-time"
          },
          "sensors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/TelemetryPoint"
              }
            },
            "description": "Keyed by sensor_id"
          }
        }
      },
...
    }
  }
}